+++ ChangeLog.md
@@ -7,6 +7,7 @@
 * Added {Ronin::Binary::Struct}.
 * Added {Ronin::Binary::Hexdump::Parser}.
 * Added {Ronin::Fuzzing::Template}.
+* Added {Ronin::Fuzzing::Repeator}.
 * Added {Ronin::Wordlist.create}.
 * Added {Ronin::Wordlist#path} and {Ronin::Wordlist#words}.
 * Added {Ronin::Wordlist#save}.